+++ src/main/java/com/example/diplom/services/OrderLineService.java
@@ -4,6 +4,7 @@ import com.example.diplom.dto.OrderLineDto;
 import com.example.diplom.entities.ApplianceEntity;
 import com.example.diplom.entities.OrderEntity;
 import com.example.diplom.entities.OrderLine;
+import com.example.diplom.exceptions.PrivateException;
 import com.example.diplom.repositories.ApplianceRepo;
 import com.example.diplom.repositories.OrderLineRepo;
 import com.example.diplom.repositories.OrderRepo;
@@ -17,17 +18,24 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class OrderLineService {
 
+    private final String NOT_BE_NULL = "Id must be greater than null";
+    private final String NO_LINE = "No such line";
+
     private final ApplianceRepo applianceRepo;
     private final OrderLineRepo orderLineRepo;
     private final OrderRepo orderRepo;
 
     public void init(Long orderId, Long applianceId) {
-        OrderLine orderLine = new OrderLine();
-        ApplianceEntity applianceEntity = applianceRepo.findById(applianceId).orElseThrow(() -> new RuntimeException("No such appliance"));
-        orderLine.setApplianceEntity(applianceEntity);
-        OrderEntity orderEntity = orderRepo.findById(orderId).orElseThrow(() -> new RuntimeException("No such order"));
-        orderLine.setOrderEntity(orderEntity);
-        orderLineRepo.save(orderLine);
+        if (orderId <= 0 || applianceId <= 0) {
+            throw new PrivateException(NOT_BE_NULL);
+        } else {
+            OrderLine orderLine = new OrderLine();
+            ApplianceEntity applianceEntity = applianceRepo.findById(applianceId).orElseThrow(() -> new RuntimeException("No such appliance"));
+            orderLine.setApplianceEntity(applianceEntity);
+            OrderEntity orderEntity = orderRepo.findById(orderId).orElseThrow(() -> new RuntimeException("No such order"));
+            orderLine.setOrderEntity(orderEntity);
+            orderLineRepo.save(orderLine);
+        }
     }
 
     public List<OrderLine> getAllOrderLines() {
@@ -35,7 +43,11 @@ public class OrderLineService {
     }
 
     public Optional<OrderLine> getById(Long numberOfLine) {
-        return orderLineRepo.findById(numberOfLine);
+        if (numberOfLine <= 0) {
+            throw new PrivateException(NOT_BE_NULL);
+        } else {
+            return orderLineRepo.findById(numberOfLine);
+        }
     }
 
     public void create(Long orderId, Long applianceId) {
@@ -43,12 +55,16 @@ public class OrderLineService {
                 applianceRepo.findById(applianceId).isPresent()) {
             init(orderId, applianceId);
         } else {
-            throw new RuntimeException("enter the correct values");
+            throw new PrivateException("Enter the correct values");
         }
     }
 
     public void deleteOrderLine(Long numberOfLine) {
-        orderLineRepo.findById(numberOfLine).orElseThrow(() -> new RuntimeException("No such line"));
-        orderLineRepo.deleteById(numberOfLine);
+        if (numberOfLine <= 0) {
+            throw new PrivateException(NOT_BE_NULL);
+        } else {
+            orderLineRepo.findById(numberOfLine).orElseThrow(() -> new PrivateException(NO_LINE));
+            orderLineRepo.deleteById(numberOfLine);
+        }
     }
 }
